+++ libraries/common/src/test/java/androidx/media3/common/audio/SonicTest.java
@@ -18,6 +18,7 @@ package androidx.media3.common.audio;
 import static androidx.media3.common.audio.Sonic.calculateAccumulatedTruncationErrorForResampling;
 import static androidx.media3.common.audio.Sonic.getExpectedFrameCountAfterProcessorApplied;
 import static androidx.media3.common.audio.Sonic.getExpectedInputFrameCountForOutputFrameCount;
+import static androidx.media3.test.utils.TestUtil.getPeriodicSamplesBuffer;
 import static com.google.common.truth.Truth.assertThat;
 
 import androidx.test.ext.junit.runners.AndroidJUnit4;
@@ -110,6 +111,53 @@ public class SonicTest {
     sonic.getOutput(outputBuffer);
 
     assertThat(outputBuffer.array()).isEqualTo(new short[] {0, 4, 8});
+  }
+
+  @Test
+  public void queueEndOfStream_withOutputCountUnderflow_setsNonNegativeOutputSize() {
+    // For speed ranges [0.5; 1) and (1; 1.5], Sonic might need to copy more input frames onto its
+    // output buffer than are available in the input buffer. Sonic keeps track of this "borrowed
+    // frames" number in #remainingInputToCopyFrameCount. When we call #queueEndOfStream(), then
+    // Sonic outputs a final number of frames based roughly on pendingOutputFrameCount +
+    // (inputFrameCount - remainingInputToCopyFrameCount) / speed + remainingInputToCopyFrameCount,
+    // which could result in a negative number if inputFrameCount < remainingInputToCopyFrameCount
+    // and 0.5 <= speed < 1. #getOutputSize() should still always return a non-negative number.
+    ShortBuffer inputBuffer =
+        getPeriodicSamplesBuffer(/* sampleCount= */ 1700, /* period= */ 192).asShortBuffer();
+    Sonic sonic =
+        new Sonic(
+            /* inputSampleRateHz= */ 48000,
+            /* channelCount= */ 1,
+            /* speed= */ 0.95f,
+            /* pitch= */ 1,
+            /* outputSampleRateHz= */ 48000);
+
+    sonic.queueInput(inputBuffer);
+    ShortBuffer outputBuffer = ShortBuffer.allocate(sonic.getOutputSize() / 2);
+    // Drain output, so that pending output frame count is 0.
+    sonic.getOutput(outputBuffer);
+    assertThat(sonic.getOutputSize()).isEqualTo(0);
+    // Queue EOS with empty pending input and output.
+    sonic.queueEndOfStream();
+
+    assertThat(sonic.getOutputSize()).isEqualTo(0);
+  }
+
+  @Test
+  public void queueEndOfStream_withNoInput_setsNonNegativeOutputSize() {
+    Sonic sonic =
+        new Sonic(
+            /* inputSampleRateHz= */ 48000,
+            /* channelCount= */ 1,
+            /* speed= */ 0.95f,
+            /* pitch= */ 1,
+            /* outputSampleRateHz= */ 48000);
+    ShortBuffer outputBuffer = ShortBuffer.allocate(sonic.getOutputSize() / 2);
+
+    sonic.getOutput(outputBuffer);
+    sonic.queueEndOfStream();
+
+    assertThat(sonic.getOutputSize()).isAtLeast(0);
   }
 
   @Test