+++ libraries/common/src/test/java/androidx/media3/common/audio/SpeedChangingAudioProcessorTest.java
@@ -17,6 +17,7 @@ package androidx.media3.common.audio;
 
 import static androidx.media3.common.audio.AudioProcessor.EMPTY_BUFFER;
 import static androidx.media3.common.util.Assertions.checkState;
+import static androidx.media3.test.utils.TestUtil.getNonRandomByteBuffer;
 import static com.google.common.truth.Truth.assertThat;
 import static org.junit.Assert.assertThrows;
 
@@ -46,12 +47,14 @@ public class SpeedChangingAudioProcessorTest {
             AUDIO_FORMAT, /* frameCounts= */ new int[] {5}, /* speeds= */ new float[] {1});
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
-    ByteBuffer inputBuffer = getInputBuffer(/* frameCount= */ 5);
+    ByteBuffer inputBuffer =
+        getNonRandomByteBuffer(/* frameCount= */ 5, AUDIO_FORMAT.bytesPerFrame);
 
     speedChangingAudioProcessor.queueInput(inputBuffer);
 
     inputBuffer.rewind();
-    assertThat(inputBuffer).isEqualTo(getInputBuffer(/* frameCount= */ 5));
+    assertThat(inputBuffer)
+        .isEqualTo(getNonRandomByteBuffer(/* frameCount= */ 5, AUDIO_FORMAT.bytesPerFrame));
   }
 
   @Test
@@ -61,12 +64,14 @@ public class SpeedChangingAudioProcessorTest {
             AUDIO_FORMAT, /* frameCounts= */ new int[] {5}, /* speeds= */ new float[] {2});
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
-    ByteBuffer inputBuffer = getInputBuffer(/* frameCount= */ 5);
+    ByteBuffer inputBuffer =
+        getNonRandomByteBuffer(/* frameCount= */ 5, AUDIO_FORMAT.bytesPerFrame);
 
     speedChangingAudioProcessor.queueInput(inputBuffer);
 
     inputBuffer.rewind();
-    assertThat(inputBuffer).isEqualTo(getInputBuffer(/* frameCount= */ 5));
+    assertThat(inputBuffer)
+        .isEqualTo(getNonRandomByteBuffer(/* frameCount= */ 5, AUDIO_FORMAT.bytesPerFrame));
   }
 
   @Test
@@ -76,7 +81,8 @@ public class SpeedChangingAudioProcessorTest {
             AUDIO_FORMAT, /* frameCounts= */ new int[] {5}, /* speeds= */ new float[] {1});
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
-    ByteBuffer inputBuffer = getInputBuffer(/* frameCount= */ 5);
+    ByteBuffer inputBuffer =
+        getNonRandomByteBuffer(/* frameCount= */ 5, AUDIO_FORMAT.bytesPerFrame);
 
     speedChangingAudioProcessor.queueInput(inputBuffer);
     speedChangingAudioProcessor.queueEndOfStream();
@@ -93,7 +99,8 @@ public class SpeedChangingAudioProcessorTest {
             AUDIO_FORMAT, /* frameCounts= */ new int[] {5}, /* speeds= */ new float[] {2});
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
-    ByteBuffer inputBuffer = getInputBuffer(/* frameCount= */ 5);
+    ByteBuffer inputBuffer =
+        getNonRandomByteBuffer(/* frameCount= */ 5, AUDIO_FORMAT.bytesPerFrame);
 
     speedChangingAudioProcessor.queueInput(inputBuffer);
     speedChangingAudioProcessor.queueEndOfStream();
@@ -111,7 +118,8 @@ public class SpeedChangingAudioProcessorTest {
             AUDIO_FORMAT, /* frameCounts= */ new int[] {5, 5}, /* speeds= */ new float[] {2, 1});
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
-    ByteBuffer inputBuffer = getInputBuffer(/* frameCount= */ 5);
+    ByteBuffer inputBuffer =
+        getNonRandomByteBuffer(/* frameCount= */ 5, AUDIO_FORMAT.bytesPerFrame);
 
     speedChangingAudioProcessor.queueInput(inputBuffer);
     inputBuffer.rewind();
@@ -131,7 +139,8 @@ public class SpeedChangingAudioProcessorTest {
             AUDIO_FORMAT, /* frameCounts= */ new int[] {5, 5}, /* speeds= */ new float[] {1, 2});
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
-    ByteBuffer inputBuffer = getInputBuffer(/* frameCount= */ 5);
+    ByteBuffer inputBuffer =
+        getNonRandomByteBuffer(/* frameCount= */ 5, AUDIO_FORMAT.bytesPerFrame);
 
     speedChangingAudioProcessor.queueInput(inputBuffer);
     inputBuffer.rewind();
@@ -159,7 +168,8 @@ public class SpeedChangingAudioProcessorTest {
             AUDIO_FORMAT, /* frameCounts= */ new int[] {5, 5}, /* speeds= */ new float[] {3, 2});
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
-    ByteBuffer inputBuffer = getInputBuffer(/* frameCount= */ 5);
+    ByteBuffer inputBuffer =
+        getNonRandomByteBuffer(/* frameCount= */ 5, AUDIO_FORMAT.bytesPerFrame);
 
     speedChangingAudioProcessor.queueInput(inputBuffer);
     inputBuffer.rewind();
@@ -187,7 +197,8 @@ public class SpeedChangingAudioProcessorTest {
             AUDIO_FORMAT, /* frameCounts= */ new int[] {5, 5}, /* speeds= */ new float[] {2, 3});
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
-    ByteBuffer inputBuffer = getInputBuffer(/* frameCount= */ 5);
+    ByteBuffer inputBuffer =
+        getNonRandomByteBuffer(/* frameCount= */ 5, AUDIO_FORMAT.bytesPerFrame);
 
     speedChangingAudioProcessor.queueInput(inputBuffer);
     ByteBuffer outputBuffer = getAudioProcessorOutput(speedChangingAudioProcessor);
@@ -214,7 +225,8 @@ public class SpeedChangingAudioProcessorTest {
             /* speeds= */ new float[] {1, 2});
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
-    ByteBuffer inputBuffer = getInputBuffer(/* frameCount= */ 5);
+    ByteBuffer inputBuffer =
+        getNonRandomByteBuffer(/* frameCount= */ 5, AUDIO_FORMAT.bytesPerFrame);
     int inputBufferLimit = inputBuffer.limit();
 
     speedChangingAudioProcessor.queueInput(inputBuffer);
@@ -233,7 +245,8 @@ public class SpeedChangingAudioProcessorTest {
             /* speeds= */ new float[] {1, 2});
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
-    ByteBuffer inputBuffer = getInputBuffer(/* frameCount= */ 5);
+    ByteBuffer inputBuffer =
+        getNonRandomByteBuffer(/* frameCount= */ 5, AUDIO_FORMAT.bytesPerFrame);
     int inputBufferLimit = inputBuffer.limit();
 
     speedChangingAudioProcessor.queueInput(inputBuffer);
@@ -252,7 +265,8 @@ public class SpeedChangingAudioProcessorTest {
             /* speeds= */ new float[] {1, 2});
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
-    ByteBuffer inputBuffer = getInputBuffer(/* frameCount= */ 5);
+    ByteBuffer inputBuffer =
+        getNonRandomByteBuffer(/* frameCount= */ 5, AUDIO_FORMAT.bytesPerFrame);
 
     // SpeedChangingAudioProcessor only queues samples until the next speed change.
     while (inputBuffer.hasRemaining()) {
@@ -276,7 +290,8 @@ public class SpeedChangingAudioProcessorTest {
             AUDIO_FORMAT, /* frameCounts= */ new int[] {5, 5}, /* speeds= */ new float[] {2, 1});
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
-    ByteBuffer inputBuffer = getInputBuffer(/* frameCount= */ 5);
+    ByteBuffer inputBuffer =
+        getNonRandomByteBuffer(/* frameCount= */ 5, AUDIO_FORMAT.bytesPerFrame);
 
     speedChangingAudioProcessor.queueInput(inputBuffer);
     inputBuffer.rewind();
@@ -295,7 +310,8 @@ public class SpeedChangingAudioProcessorTest {
             AUDIO_FORMAT, /* frameCounts= */ new int[] {5, 5}, /* speeds= */ new float[] {1, 2});
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
-    ByteBuffer inputBuffer = getInputBuffer(/* frameCount= */ 5);
+    ByteBuffer inputBuffer =
+        getNonRandomByteBuffer(/* frameCount= */ 5, AUDIO_FORMAT.bytesPerFrame);
 
     speedChangingAudioProcessor.queueInput(inputBuffer);
     inputBuffer.rewind();
@@ -314,7 +330,8 @@ public class SpeedChangingAudioProcessorTest {
             AUDIO_FORMAT, /* frameCounts= */ new int[] {5}, /* speeds= */ new float[] {1});
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
-    ByteBuffer inputBuffer = getInputBuffer(/* frameCount= */ 5);
+    ByteBuffer inputBuffer =
+        getNonRandomByteBuffer(/* frameCount= */ 5, AUDIO_FORMAT.bytesPerFrame);
 
     speedChangingAudioProcessor.queueInput(inputBuffer);
     speedChangingAudioProcessor.queueEndOfStream();
@@ -330,7 +347,8 @@ public class SpeedChangingAudioProcessorTest {
             AUDIO_FORMAT, /* frameCounts= */ new int[] {5}, /* speeds= */ new float[] {2});
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
-    ByteBuffer inputBuffer = getInputBuffer(/* frameCount= */ 5);
+    ByteBuffer inputBuffer =
+        getNonRandomByteBuffer(/* frameCount= */ 5, AUDIO_FORMAT.bytesPerFrame);
 
     speedChangingAudioProcessor.queueInput(inputBuffer);
     speedChangingAudioProcessor.queueEndOfStream();
@@ -358,7 +376,8 @@ public class SpeedChangingAudioProcessorTest {
             AUDIO_FORMAT, /* frameCounts= */ new int[] {5}, /* speeds= */ new float[] {1});
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
-    ByteBuffer inputBuffer = getInputBuffer(/* frameCount= */ 5);
+    ByteBuffer inputBuffer =
+        getNonRandomByteBuffer(/* frameCount= */ 5, AUDIO_FORMAT.bytesPerFrame);
 
     speedChangingAudioProcessor.queueInput(inputBuffer);
     getAudioProcessorOutput(speedChangingAudioProcessor);
@@ -373,7 +392,8 @@ public class SpeedChangingAudioProcessorTest {
             AUDIO_FORMAT, /* frameCounts= */ new int[] {5}, /* speeds= */ new float[] {2});
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
-    ByteBuffer inputBuffer = getInputBuffer(/* frameCount= */ 5);
+    ByteBuffer inputBuffer =
+        getNonRandomByteBuffer(/* frameCount= */ 5, AUDIO_FORMAT.bytesPerFrame);
 
     speedChangingAudioProcessor.queueInput(inputBuffer);
     getAudioProcessorOutput(speedChangingAudioProcessor);
@@ -390,7 +410,8 @@ public class SpeedChangingAudioProcessorTest {
             AUDIO_FORMAT, /* frameCounts= */ new int[] {5, 5}, /* speeds= */ new float[] {2, 1});
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
-    ByteBuffer inputBuffer = getInputBuffer(/* frameCount= */ 5);
+    ByteBuffer inputBuffer =
+        getNonRandomByteBuffer(/* frameCount= */ 5, AUDIO_FORMAT.bytesPerFrame);
 
     speedChangingAudioProcessor.getSpeedAdjustedTimeAsync(
         /* inputTimeUs= */ 50L, outputTimesUs::add);
@@ -421,7 +442,8 @@ public class SpeedChangingAudioProcessorTest {
             /* speeds= */ new float[] {2, 1, 3});
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
-    ByteBuffer inputBuffer = getInputBuffer(/* frameCount= */ 5);
+    ByteBuffer inputBuffer =
+        getNonRandomByteBuffer(/* frameCount= */ 5, AUDIO_FORMAT.bytesPerFrame);
     // Use the audio processor before a flush
     speedChangingAudioProcessor.queueInput(inputBuffer);
     getAudioProcessorOutput(speedChangingAudioProcessor);
@@ -458,7 +480,8 @@ public class SpeedChangingAudioProcessorTest {
             AUDIO_FORMAT, /* frameCounts= */ new int[] {5, 5}, /* speeds= */ new float[] {2, 1});
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
-    ByteBuffer inputBuffer = getInputBuffer(/* frameCount= */ 3);
+    ByteBuffer inputBuffer =
+        getNonRandomByteBuffer(/* frameCount= */ 3, AUDIO_FORMAT.bytesPerFrame);
 
     speedChangingAudioProcessor.getSpeedAdjustedTimeAsync(
         /* inputTimeUs= */ 300L, outputTimesUs::add);
@@ -487,7 +510,8 @@ public class SpeedChangingAudioProcessorTest {
             AUDIO_FORMAT, /* frameCounts= */ new int[] {5, 5}, /* speeds= */ new float[] {2, 1});
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
-    ByteBuffer inputBuffer = getInputBuffer(/* frameCount= */ 5);
+    ByteBuffer inputBuffer =
+        getNonRandomByteBuffer(/* frameCount= */ 5, AUDIO_FORMAT.bytesPerFrame);
     speedChangingAudioProcessor.queueInput(inputBuffer);
     getAudioProcessorOutput(speedChangingAudioProcessor);
     inputBuffer.rewind();
@@ -513,7 +537,8 @@ public class SpeedChangingAudioProcessorTest {
             /* speeds= */ new float[] {2, 1, 5, 2});
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
-    ByteBuffer inputBuffer = getInputBuffer(/* frameCount= */ 441 * 4);
+    ByteBuffer inputBuffer =
+        getNonRandomByteBuffer(/* frameCount= */ 441 * 4, AUDIO_FORMAT.bytesPerFrame);
     while (inputBuffer.position() < inputBuffer.limit()) {
       speedChangingAudioProcessor.queueInput(inputBuffer);
     }
@@ -552,7 +577,7 @@ public class SpeedChangingAudioProcessorTest {
             /* speeds= */ new float[] {2, 4, 2}); // 500, 250, 500 = 1250
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
-    ByteBuffer input = getInputBuffer(1000);
+    ByteBuffer input = getNonRandomByteBuffer(1000, AUDIO_FORMAT.bytesPerFrame);
 
     speedChangingAudioProcessor.queueInput(input);
     outputFrameCount +=
@@ -587,7 +612,7 @@ public class SpeedChangingAudioProcessorTest {
             /* speeds= */ new float[] {2, 3, 8, 4});
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
-    ByteBuffer input = getInputBuffer(12);
+    ByteBuffer input = getNonRandomByteBuffer(12, AUDIO_FORMAT.bytesPerFrame);
 
     while (input.hasRemaining()) {
       speedChangingAudioProcessor.queueInput(input);
@@ -614,7 +639,7 @@ public class SpeedChangingAudioProcessorTest {
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
     // 1500 input frames falls in the middle of the 2x region.
-    ByteBuffer input = getInputBuffer(1500);
+    ByteBuffer input = getNonRandomByteBuffer(1500, AUDIO_FORMAT.bytesPerFrame);
     int outputFrameCount = 0;
 
     while (input.hasRemaining()) {
@@ -653,7 +678,7 @@ public class SpeedChangingAudioProcessorTest {
     SpeedChangingAudioProcessor speedChangingAudioProcessor =
         getConfiguredSpeedChangingAudioProcessor(speedProvider);
     // 1500 input frames falls in the middle of the 2x region.
-    ByteBuffer input = getInputBuffer(1500);
+    ByteBuffer input = getNonRandomByteBuffer(1500, AUDIO_FORMAT.bytesPerFrame);
     int outputFrameCount = 0;
 
     while (input.hasRemaining()) {
@@ -808,16 +833,6 @@ public class SpeedChangingAudioProcessorTest {
     speedChangingAudioProcessor.configure(AUDIO_FORMAT);
     speedChangingAudioProcessor.flush();
     return speedChangingAudioProcessor;
-  }
-
-  private static ByteBuffer getInputBuffer(int frameCount) {
-    int bufferSize = frameCount * AUDIO_FORMAT.bytesPerFrame;
-    ByteBuffer buffer = ByteBuffer.allocateDirect(bufferSize).order(ByteOrder.nativeOrder());
-    for (int i = 0; i < bufferSize; i++) {
-      buffer.put((byte) (i % (Byte.MAX_VALUE + 1)));
-    }
-    buffer.rewind();
-    return buffer;
   }
 
   private static ByteBuffer getAudioProcessorOutput(AudioProcessor audioProcessor) {